python/infoscreen/requirements.txt
ics
fmiopendata
eccodes
bs4
request
python-dotenv
